+++ changelog.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Add [changelog.md](changelog.md) and [readme.md](readme.md) to each build of a release.
 - Add for each target programming language a new new command-line argument (e.g. `--java`, `--c` or `--go`).
 - Add pipe functionality and the related command-line argument (`--pipe` or `-p`).
+- Add test class `Go` in `tests/language/Go.py` to test all implementations for the target programming language Go. 
  
 ### Changed